// src/api/category/messages/category.messages.ts
export enum CategoryMessages {
  CATEGORY_CREATED = 'Categoría creada con éxito',
  SUBCATEGORY_CREATED = 'Subcategoría creada con éxito',
  SUBCATEGORY_UPDATED = 'Subcategoría actualizada con éxito',
  CATEGORY_UPDATED = 'Categoría actualizada con éxito',
  CATEGORY_ACTIVED = 'Categoria o subcategoria activada con éxito',
  CATEGORY_DESACTIVED = 'Categoria o subcategoria desactivada con éxito',
  CATEGORY_REMOVED = 'Categoría eliminada con éxito',
  SUBCATEGORY_REMOVED = 'Subcategoría eliminada con éxito',
  CATEGORY_RESTORED = 'Categoría o subcategoria restaurada con éxito',
  SUBCATEGORY_RESTORED = 'Subcategoria restaurada con éxito',
}